tests/requirements.txt
newspaper3k==0.2.8
lxml==5.1.0
